product action button component
