--- src/components/ProductDetails/ProductActionButton/ProductActionButton.jsx
@@ -1,13 +1,0 @@
-import "./ProductActionButton.scss";
-
-function ProductActionButton(props) {
-  const buttonText = props.buttonText;
-  return (
-    <div className="product-action-button-wrapper">
-      <button className="product-action-button-wrapper__button">
-        {buttonText}
-      </button>
-    </div>
-  );
-}
-export default ProductActionButton;
